Implemented the matrix `Element` type (with better conversion from `float`).
This type is a subclass of the `decimal.Decimal` type.
Hence, it has much better support across different real number types.
It retains integers as they're and the availabilty of big integers.
It has better floating-point support than `float`, with extended precision.

NOTE: Had to implement support for inter-operations with `float` instances.